refactor: renaming entities
