+++ socket/src/main/scala/io/github/positionpal/server/routes/Routes.scala
@@ -4,8 +4,8 @@ import akka.actor.typed.ActorSystem
 import akka.http.scaladsl.server.Directives.*
 import akka.http.scaladsl.model.*
 import akka.http.scaladsl.server.Route
-import io.github.positionpal.server.ws.Handlers.websocketHandler
-import io.github.positionpal.entity.WebSocketActor.{Commands, incomingHandler}
+import io.github.positionpal.server.ws.WebSocketHandlers.websocketHandler
+import io.github.positionpal.entity.Handler.{Commands, incomingHandler}
 
 
 object Routes:
